Important bug fix related to MPI data type.
Sending size_t with MPI is dangerous as it can represent many types:
http://stackoverflow.com/questions/40807833/sending-size-t-type-data-with-mpi

So we use int for all dimension variables for safety.
Although the array size is thus less flexible.

NetCDF nc_inq_dimlen returns size_t. Simply convert it back to int.

+++ Parallel_Algorithm/OpenMP/Kmean_omp.c
@@ -18,7 +18,7 @@ int main() {
     ----------------  Initialization ---------------------
     ======================================================
     */
-    size_t N_samples,N_features,N_clusters,N_repeat;
+    int N_samples,N_features,N_clusters,N_repeat;
     //i for samples; j for features; k for clusters (typically)
     int i,j,k;
     int k_best,initial_idx;
