+++ Parallel_Algorithm/shared/ncdf_util.h
@@ -2,8 +2,8 @@
 #define NCDF_UTIL_H
 
 int readX(char* FILE_NAME, float*** p_X,int*** p_GUESS,
-          size_t* p_N_samples,size_t* p_N_features,
-          size_t* p_N_clusters,size_t* p_N_repeat );
+          int* p_N_samples,int* p_N_features,
+          int* p_N_clusters,int* p_N_repeat );
 
 int writeY(char* FILE_NAME, int* labels, float inert);
